age_core: Expose primitives for AEAD and HKDF
This enables reuse of these common primitives within the upcoming plugin
system, alongside the core parsers and serializers.
